going to test usecase

+++ pkg/repository/interface/user_repo.go
@@ -3,6 +3,6 @@ package interfaces
 import "github.com/ashiqsabith123/auth-svc/pkg/domain"
 
 type UserRepo interface {
-	FindUser(phone int) (bool, error)
+	FindUser(phone int64) (bool, error)
 	CreateUser(user domain.User) error
 }
